+++ src/ngdexie.utils.js
@@ -1,21 +1,21 @@
 (function () {
     'use strict';
-
     /**
      * Create ngdexie.utils module
      */
     angular.module('ngdexie.utils', ['ngdexie.core']);
-
     /**
      * Create ngDexieUtils factory
      */
     angular.module('ngdexie.utils')
             .factory("ngDexieUtils", ngDexieUtils);
 
+    /*@ngInject*/
     function ngDexieUtils() {
 
         return {
-            deepClone: deepClone
+            deepClone: deepClone,
+            debounce: debounce
         };
 
         /**
@@ -30,6 +30,34 @@
             }
             return value;
         }
-    }
 
+        /**
+         * Protect a function for being called to rapidly
+         * @param {type} func
+         * @param {type} wait
+         * @param {type} immediate
+         * @returns {Function}
+         */
+        function debounce(func, wait, immediate) {
+            console.log("debounce 1");
+            var timeout;
+            return function () {
+                console.log("debounce 2");
+                var context = this, args = arguments;
+                var later = function () {
+                    console.log("debounce later");
+                    timeout = null;
+                    if (!immediate) {
+                        func.apply(context, args);
+                    }
+                };
+                var callNow = immediate && !timeout;
+                clearTimeout(timeout);
+                timeout = setTimeout(later, wait);
+                if (callNow) {
+                    func.apply(context, args);
+                }
+            };
+        }
+    }
 })();